Move control API client to device node
Moved the code for accessing the Sony audio control API from the
controller to the device node.
